requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
decorator==5.1.1
distlib==0.3.8
Django==5.1.6
django-cors-headers==4.6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
ipdb==0.13.13
ipython==8.22.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.7
numpy==1.25.1
opencv-python==4.8.0.74
openpyxl==3.1.5
orjson==3.10.6
packaging==24.2
pandas==2.0.3
parso==0.8.3
platformdirs==4.2.1
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.17.2
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
traitlets==5.14.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.26.1
virtualenvwrapper-win==1.2.7
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
